# trireg-java-client-api/target/generated-sources/openapi/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: example@example.com
    name: Trifork A/S
    url: https://github.com/trifork/trireg2-api
  description: An OpenAPI specification for time registrations to be used by a user
  license:
    name: MIT License
    url: https://github.com/trifork/trireg2-api/LICENSE.md
  title: TriReg2 User API
  version: 0.2.0
servers:
- description: Test server (uses test data)
  url: https://trireg2.tcs.trifork.dev
- description: Local development server
  url: http://0.0.0.0:8080
security:
- openId: []
- basicAuth: []
...
paths:
...
  /time-registration:
    delete:
      description: Deletes the specified time registration
      operationId: deleteTimeRegistration
      parameters:
      - description: The id of the time registration to delete
        explode: true
        in: query
        name: timeRegistrationId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                success-deleted:
                  value:
                    id: 12345
                    status: The time registration was deleted successfully (Status
                      = deleted)
                success-rejected:
                  value:
                    id: 54321
                    status: The imported time registration was rejected successfully
                      (Status = rejected)
              schema:
                $ref: '#/components/schemas/DefaultDeleteResponse'
          description: deleted
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Bad delete request
        "401":
          description: JWT is missing or invalid
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Time registration not found
      tags:
      - TimeRegistration
      x-accepts: "application/json,text/plain"
    get:
      description: Returns time registrations that belongs to a user
      operationId: getTimeRegistrationsForUser
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TimeRegistrationResponse'
                type: array
          description: List of time registrations for the user
        "401":
          description: JWT is missing or invalid
      tags:
      - TimeRegistration
      x-accepts: application/json
    post:
      description: Adds a time registration for a user
      operationId: addTimeRegistrationForUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeRegistrationRequest'
        description: A JSON object containing time registration information
        required: true
      responses:
        "201":
          content:
            application/json:
              examples:
                Example#1:
                  value:
                    id: 12345
                    status: Time registration added successfully
              schema:
                $ref: '#/components/schemas/DefaultCreateResponse'
          description: Time registration added successfully
        "401":
          description: JWT is missing or invalid
      tags:
      - TimeRegistration
      x-content-type: application/json
      x-accepts: application/json
    put:
      description: Updates the specified time registration
      operationId: updateTimeRegistrationForUser
      parameters:
      - description: The id of the time registration to update
        explode: true
        in: query
        name: timeRegistrationId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeRegistrationUpdateRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                Success:
                  value:
                    id: 12345
                    status: Time registration updated successfully
              schema:
                $ref: '#/components/schemas/DefaultCreateResponse'
          description: Updated
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Bad update request
        "401":
          description: JWT is missing or invalid
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Time registration not found
      tags:
      - TimeRegistration
      x-content-type: application/json
      x-accepts: "application/json,text/plain"
  /time-registration/active-task:
    get:
      description: Returns time registrations for a user grouped by task
      operationId: getTaskTimeRegistrationsOverview
      parameters:
      - description: The date to consider when fetching active tasks and time registrations
        explode: true
        in: query
        name: date
        required: true
        schema:
          format: date
          type: string
        style: form
      - description: |
          The time period to use when deciding in which time period tasks are considered active. Defaults to 'WEEK'.
        explode: true
        in: query
        name: period
        required: false
        schema:
          $ref: '#/components/schemas/OverviewPeriod'
        style: form
      - description: "An optional collection of extra tasks to include in the response,\
          \ regardless of whether they are active or not"
        explode: true
        in: query
        name: extraTasks
        required: false
        schema:
          items:
            format: int64
            type: integer
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeRegistrationsByTaskResponse'
          description: List of time registrations for the user grouped by task
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Request contains an invalid time period value
        "401":
          description: JWT is missing or invalid
      tags:
      - TimeRegistration
      x-accepts: "application/json,text/plain"
...
  /export:
    post:
      description: Exports time registrations for a specific time period for the user
        logged in
      operationId: exportTimeRegistrationsForUser
      parameters:
      - description: The start date of the export time interval (inclusive)
        explode: true
        in: query
        name: start
        required: true
        schema:
          example: 2024-03-11
          format: date
          type: string
        style: form
      - description: The end date of the export time interval (inclusive)
        explode: true
        in: query
        name: end
        required: true
        schema:
          example: 2024-03-17
          format: date
          type: string
        style: form
      responses:
        "200":
          content:
            text/plain:
              examples:
                Example#1:
                  value: Time registrations exported successfully
              schema:
                type: string
          description: OK
        "401":
          description: JWT is missing or invalid
        "502":
          description: "While acting as a gateway or proxy, this service received\
            \ an invalid response from the upstream server."
      tags:
      - Export
      x-accepts: text/plain
...
  /import:
    post:
      description: Imports the given time registrations into the system
      operationId: importTimeRegistrations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportTimeRegistrationRequest'
        description: A JSON object containing time registration information to be
          imported
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportTimeRegistrationResponse'
          description: OK
        "401":
          description: JWT is missing or invalid
        "502":
          description: "While acting as a gateway or proxy, this service received\
            \ an invalid response from the upstream server."
      tags:
      - Import
      x-content-type: application/json
      x-accepts: application/json
components:
...
  schemas:
    TimeRegistrationResponse:
      example:
        date: 2000-01-23
        duration: PT1H30M
        timeRegistrationId: 0
        registered: 6
        description: description
        userId: 1
        taskId: 5
        status: null
        tags:
        - tagConfigurationMetadata:
            tagConfigurationDescription: tagConfigurationDescription
            tagConfigurationId: 5
            tagConfigurationName: tagConfigurationName
            valueType: STRING
            cardinality: OPTIONAL
          tagValue:
            tagId: 2
            tagValue: tagValue
        - tagConfigurationMetadata:
            tagConfigurationDescription: tagConfigurationDescription
            tagConfigurationId: 5
            tagConfigurationName: tagConfigurationName
            valueType: STRING
            cardinality: OPTIONAL
          tagValue:
            tagId: 2
            tagValue: tagValue
      properties:
        timeRegistrationId:
          format: int64
          type: integer
        registered:
          description: The registration time as Unix milliseconds timestamp
          format: int64
          type: integer
        userId:
          format: int64
          type: integer
        taskId:
          format: int64
          type: integer
        date:
          format: date
          type: string
        duration:
          description: "ISO 8601 formatted duration component <PnDTnHnMn.nS>. Each\
            \ unit is optional, but at least one must be specified"
          example: PT1H30M
          format: duration
          type: string
        status:
          $ref: '#/components/schemas/TimeRegistrationStatus'
        tags:
          items:
            $ref: '#/components/schemas/TimeRegistrationTag'
          type: array
        description:
          type: string
      required:
      - date
      - duration
      - registered
      - status
      - tags
      - timeRegistrationId
      - userId
      type: object
    TimeRegistrationUpdateRequest:
      example:
        date: 2000-01-23
        duration: PT1H30M
        description: description
        status: null
      properties:
        date:
          format: date
          type: string
        duration:
          description: "ISO 8601 formatted duration component <PnDTnHnMn.nS>. Each\
            \ unit is optional, but at least one must be specified"
          example: PT1H30M
          format: duration
          type: string
        status:
          $ref: '#/components/schemas/TimeRegistrationStatus'
        description:
          type: string
      type: object
    DefaultCreateResponse:
      description: A response to a create operation.
      example:
        id: 12345
        status: The data was created successfully
      properties:
        id:
          description: The id of the created entity.
          format: int64
          type: integer
        status:
          description: A description of the status of the create operation.
          type: string
      required:
      - id
      - status
      type: object
    TimeRegistrationRequest:
      example:
        date: 2000-01-23
        duration: PT1H30M
        description: description
        taskId: 0
        tags:
        - tagConfigurationId: 6
          value: value
        - tagConfigurationId: 6
          value: value
      properties:
        taskId:
          description: |
            Associates the time registration with a task.
            If not specified, then the `description` field must be a non-null value
          format: int64
          type: integer
        date:
          format: date
          type: string
        duration:
          description: "ISO 8601 formatted duration component <PnDTnHnMn.nS>. Each\
            \ unit is optional, but at least one must be specified"
          example: PT1H30M
          format: duration
          type: string
        tags:
          items:
            $ref: '#/components/schemas/TagTimeRegistrationRequest'
          type: array
        description:
          type: string
      required:
      - date
      - duration
      type: object
    DefaultDeleteResponse:
      description: A response to a delete operation.
      example:
        id: 12345
        status: The data was deleted successfully
      properties:
        id:
          format: int64
          type: integer
        status:
          description: A description of the status of the delete operation.
          type: string
      required:
      - id
      - status
      type: object
    OverviewPeriod:
      description: |
        A period of time that an overview is covering Valid values:
            * `day` - Covers a specific date
            * `week` - Covers the whole week (Monday to Sunday) of a specific date
            * `month` - Covers the whole month of a specific date
      enum:
      - DAY
      - WEEK
      - MONTH
      type: string
...
    TimeRegistrationTag:
      description: "Composite object with information about a potential tag for a\
        \ time registration. This consists both of tags that has\nbeen set, and tags\
        \ that can potentially be set. Contains metadata describing the tag and the\
        \ current tag value (if \nany). The 'tagValue' is null if no tag value has\
        \ been set for the specific configuration.\n"
      example:
        tagConfigurationMetadata:
          tagConfigurationDescription: tagConfigurationDescription
          tagConfigurationId: 5
          tagConfigurationName: tagConfigurationName
          valueType: STRING
          cardinality: OPTIONAL
        tagRegistration:
          tagId: 2
          tagRegistration: tagRegistration
      properties:
        tagConfigurationMetadata:
          $ref: '#/components/schemas/TagConfigurationMetadata'
        tagRegistration:
          $ref: '#/components/schemas/TagRegistration'
      required:
      - tagConfigurationMetadata
      type: object
    TagRegistration:
      description: An existing tag with some value
      example:
        tagId: 2
        tagValue: tagValue
      properties:
        tagId:
          description: The id of this tag
          format: int64
          type: integer
        tagValue:
          description: "The value of this tag. This is a valid value corresponding\
            \ to the tag configuration. For example, if the tag \nconfiguration specifies\
            \ that the tag is an integer, then the value is a valid integer.\n"
          type: string
      required:
      - tagId
      - tagValue
      type: object
    TagTimeRegistrationRequest:
      description: |
        A request to set a tag value for a tag configuration for a time registration. Administrators configure the rules for
        tags for time registrations. This includes the format of the tag value, and whether tags are mandatory or optional.
        When setting a tag value for a time registration, the request must follow these rules.
      example:
        tagConfigurationId: 6
        value: value
      properties:
        tagConfigurationId:
          description: "The id of the tag configuration to set a value for. This must\
            \ be a valid tag configuration for the time \nregistration.\n"
          format: int64
          type: integer
        value:
          description: |-
            The value to set for the tag configuration. This is a string, but the format of the string is dependent on the tag
            configuration. For example, if the configuration specifies that the tag is an integer, then the string value must
            be a valid integer.
          type: string
      required:
      - tagConfigurationId
      - value
      type: object
    ImportTimeRegistrationRequest:
      items:
        $ref: '#/components/schemas/ImportTimeRegistrationRequest_inner'
      type: array
    ImportTimeRegistrationResponse:
      items:
        $ref: '#/components/schemas/ImportTimeRegistrationResponse_inner'
      type: array
...
    TimeRegistrationStatus:
      description: "The status of a time registration Valid values: * `VALID` - The\
        \ time registration is valid * `PENDING` - The time registration is valid\
        \ as such, but is awaiting the approval of the user. If this happens, it will\
        \ become 'valid'. If it is rejected, it will become 'rejected'. * `INVALID`\
        \ - Something is wrong with the time registration, it will need correction\
        \ before it can become valid * `DELETED` - The time registration has existed\
        \ at some point, but has now been deleted * `REJECTED` - The time registration\
        \ has been rejected by the user"
      enum:
      - VALID
      - PENDING
      - INVALID
      - DELETED
      - REJECTED
      type: string
...
    TagConfigurationMetadata:
      description: |
        Metadata for a tag configuration in the context of a specific time registration. This can differ based on the task the
        time registration is associated to, so a tag that is optional for one time registration can be mandatory for another.
      example:
        tagConfigurationDescription: tagConfigurationDescription
        tagConfigurationId: 5
        tagConfigurationName: tagConfigurationName
        valueType: STRING
        cardinality: OPTIONAL
      properties:
        tagConfigurationId:
          description: An identifier for this tag configuration
          format: int64
          type: integer
        tagConfigurationName:
          description: The name of the tag configuration
          type: string
        tagConfigurationDescription:
          description: An optional description of the tag configuration
          type: string
        cardinality:
          description: |
            The status of a time registration Valid values: * `OPTIONAL` - The tag can be added to this time registration, but it is not required. * `MANDATORY` - The tag must be added to this time registration. * `MANAGED` - The tag is managed by administrator configuration. It is read-only and can't be changed.
          enum:
          - OPTIONAL
          - MANDATORY
          - MANAGED
          type: string
        valueType:
          description: "The value type of this configuration Valid values: * `STRING`\
            \ - The tag value can be any string * `BOOL` - The tag value must be a\
            \ boolean on the form 'true' or 'false' (case insensitive) * `INT` - The\
            \ tag value must be a valid 32-bit integer"
          enum:
          - STRING
          - BOOL
          - INT
          type: string
      required:
      - cardinality
      - tagConfigurationId
      - tagConfigurationName
      - valueType
      type: object
    ImportTimeRegistrationStatus:
      description: "The status of importing a time registration Valid values: * `VALID`\
        \ - The time registration is valid and has been imported successfully * `PENDING`\
        \ - The time registration is valid, but is awaiting the approval of the user.\
        \ Used when a user imports data for another user. * `FAILED` - Something is\
        \ wrong with the time registration, it has not been imported."
      enum:
      - VALID
      - PENDING
      - FAILED
      type: string
...
    ImportTimeRegistrationRequest_inner:
      example:
        date: 2000-01-23
        duration: PT1H30M
        description: description
        userId: 0
        taskId: 6
        tags:
        - tagConfigurationId: 6
          value: value
        - tagConfigurationId: 6
          value: value
      properties:
        userId:
          format: int64
          type: integer
        taskId:
          description: |
            Associates the time registration with a task.
            If not specified, then the `description` field must be a non-null value
          format: int64
          type: integer
        date:
          format: date
          type: string
        duration:
          description: "ISO 8601 formatted duration component <PnDTnHnMn.nS>. Each\
            \ unit is optional, but at least one must be specified"
          example: PT1H30M
          format: duration
          type: string
        tags:
          items:
            $ref: '#/components/schemas/TagTimeRegistrationRequest'
          type: array
        description:
          type: string
      required:
      - date
      - duration
      - tags
      - userId
      type: object
    ImportTimeRegistrationResponse_inner:
      example:
        timeRegistrationId: 0
        message: message
        status: null
      properties:
        timeRegistrationId:
          format: int64
          type: integer
        status:
          $ref: '#/components/schemas/ImportTimeRegistrationStatus'
        message:
          description: Provides a description of the import status
          type: string
      required:
      - status
      - timeRegistrationId
      type: object
...
  securitySchemes:
    openId:
      openIdConnectUrl: https://oidc.hosted.trifork.com/auth/realms/trifork/.well-known/openid-configuration
      type: openIdConnect
    basicAuth:
      scheme: basic
      type: http